Update feed to handle NFT
