+++ components/VideoDetail.tsx
@@ -1,0 +1,72 @@
+import { HiVolumeOff, HiVolumeUp } from "react-icons/hi";
+import React, { useEffect, useRef, useState } from "react";
+
+import { BsFillPlayFill } from "react-icons/bs";
+import { MdOutlineCancel } from "react-icons/md";
+import post from "../pages/api/post";
+import router from "next/router";
+
+const VideoDetail = ({ url }) => {
+  const videoRef = useRef<HTMLVideoElement>(null);
+  const [playing, setPlaying] = useState(false);
+  const [isVideoMuted, setIsVideoMuted] = useState(false);
+
+  const onVideoClick = () => {
+    if (playing) {
+      videoRef?.current?.pause();
+      setPlaying(false);
+    } else {
+      videoRef?.current?.play();
+      setPlaying(true);
+    }
+  };
+
+  useEffect(() => {
+    if (videoRef?.current) {
+      videoRef.current.muted = isVideoMuted;
+    }
+  }, [isVideoMuted]);
+
+  return (
+    <>
+      <div className="relative flex-2 w-[1000px] lg:w-9/12 flex justify-center items-center bg-black">
+        <div className="absolute top-6 left-2 lg:left-6 flex gap-6 z-50">
+          <p className="cursor-pointer" onClick={() => router.back()}>
+            <MdOutlineCancel className="text-white text-[35px]" />
+          </p>
+        </div>
+        <div className="relative">
+          <div className="lg:h-[100vh] h-[60vh]">
+            <video
+              ref={videoRef}
+              loop
+              onClick={onVideoClick}
+              src={url}
+              className="h-full cursor-pointer"
+            ></video>
+          </div>
+          <div className="absolute top-[45%] left-[45%] cursor-pointer">
+            {!playing && (
+              <button onClick={onVideoClick}>
+                <BsFillPlayFill className="text-white text-6xl lg:text-8xl" />
+              </button>
+            )}
+          </div>
+        </div>
+        <div className="absolute bottom-5 lg:bottom-10 right-5 lg:right-10 cursor">
+          {isVideoMuted ? (
+            <button onClick={() => setIsVideoMuted(false)}>
+              <HiVolumeOff className="text-white text-2xl lg:text-4xl" />
+            </button>
+          ) : (
+            <button onClick={() => setIsVideoMuted(true)}>
+              <HiVolumeUp className="text-white text-2xl lg:text-4xl" />
+            </button>
+          )}
+        </div>
+      </div>
+    </>
+  );
+};
+
+export default VideoDetail;
