+++ pages/detail/[id].tsx
@@ -1,5 +1,4 @@
-import { HiVolumeOff, HiVolumeUp } from "react-icons/hi";
-import React, { useEffect, useRef, useState } from "react";
+import React, { useCallback, useState } from "react";
 import {
   ReactionBarSelector,
   ReactionCounter,
@@ -7,18 +6,18 @@ import {
 import { reactionEmojis, selectorEmojis } from "../../utils/constants";
 
 import { BASE_URL } from "../../utils";
-import { BsFillPlayFill } from "react-icons/bs";
 import Comments from "../../components/Comments";
 import { GoVerified } from "react-icons/go";
 import Image from "next/image";
 import Link from "next/link";
-import { MdOutlineCancel } from "react-icons/md";
+
 import { Video } from "../../types";
 import axios from "axios";
 import findKey from "lodash.findkey";
 import useAuthStore from "../../store/authStore";
-import { useRouter } from "next/router";
 import without from "lodash.without";
+import VideoDetail from "../../components/VideoDetail";
+import NftDetail from "../../components/NftDetail";
 
 type ReactionsObject = {
   reactionThumbsUp: string;
@@ -32,29 +31,25 @@ type ReactionsObject = {
 const Detail = ({ postDetails }: IProps) => {
   const [mappedReactions, setMappedReactions] = useState([]);
   const [post, setPost] = useState(postDetails);
-  const [playing, setPlaying] = useState(false);
-  const [isVideoMuted, setIsVideoMuted] = useState(false);
   const [usedReactions, setUsedReactions] = useState(Array<string>);
-  const videoRef = useRef<HTMLVideoElement>(null);
-  const router = useRouter();
+
   const { userProfile, allUsers }: any = useAuthStore();
 
-  const onVideoClick = () => {
-    if (playing) {
-      videoRef?.current?.pause();
-      setPlaying(false);
-    } else {
-      videoRef?.current?.play();
-      setPlaying(true);
-    }
-  };
+  useCallback(() => {
+    const counterObjects = Object.keys(reactionEmojis).flatMap((reaction) => {
+      return post.reactions[reaction]?.map((id: string) => ({
+        node: (
+          <div className="mx-1">
+            {reactionEmojis[reaction as keyof ReactionsObject]}
+          </div>
+        ),
+        label: reaction,
+        by: allUsers.find(({ _id }) => _id == id)?.userName,
+      }));
+    });
 
-  useEffect(() => {
-    if (post && videoRef?.current) {
-      videoRef.current.muted = isVideoMuted;
-    }
-    reactionsFromPost(post);
-  }, [post, isVideoMuted]);
+    setMappedReactions(counterObjects.filter((e) => e) as []);
+  }, [allUsers, post.reactions]);
 
   const reactionCounterStyle: CSS.Properties = {
     marginLeft: "5px",
@@ -118,7 +113,11 @@ const Detail = ({ postDetails }: IProps) => {
     const counterObjects = Object.keys(reactionEmojis).flatMap(
       (reactionEmoji: string) => {
         return response[reactionEmoji]?.map((reaction: ReactionsObject) => ({
-          node: <div className="mx-1">{reactionEmojis[reactionEmoji as keyof ReactionsObject]}</div>,
+          node: (
+            <div className="mx-1">
+              {reactionEmojis[reactionEmoji as keyof ReactionsObject]}
+            </div>
+          ),
           label: "smile",
           by: allUsers.find(({ _id }) => _id == reaction._ref)?.userName,
         }));
@@ -127,58 +126,15 @@ const Detail = ({ postDetails }: IProps) => {
     setMappedReactions(counterObjects.filter((e) => e) as []);
   };
 
-  const reactionsFromPost = (_post: Video) => {
-    const counterObjects = Object.keys(reactionEmojis).flatMap((reaction) => {
-      return _post.reactions[reaction]?.map((id: string) => ({
-        node: <div className="mx-1">{reactionEmojis[reaction as keyof ReactionsObject]}</div>,
-        label: reaction,
-        by: allUsers.find(({ _id }) => _id == id)?.userName,
-      }));
-    });
-
-    setMappedReactions(counterObjects.filter((e) => e) as []);
-  };
-
   if (!post) return null;
 
   return (
     <div className="flex w-full absolute left-0 top-0 bg-white flex-wrap lg:flex-nowrap">
-      <div className="relative flex-2 w-[1000px] lg:w-9/12 flex justify-center items-center bg-black">
-        <div className="absolute top-6 left-2 lg:left-6 flex gap-6 z-50">
-          <p className="cursor-pointer" onClick={() => router.back()}>
-            <MdOutlineCancel className="text-white text-[35px]" />
-          </p>
-        </div>
-        <div className="relative">
-          <div className="lg:h-[100vh] h-[60vh]">
-            <video
-              ref={videoRef}
-              loop
-              onClick={onVideoClick}
-              src={post.video.asset.url}
-              className="h-full cursor-pointer"
-            ></video>
-          </div>
-          <div className="absolute top-[45%] left-[45%] cursor-pointer">
-            {!playing && (
-              <button onClick={onVideoClick}>
-                <BsFillPlayFill className="text-white text-6xl lg:text-8xl" />
-              </button>
-            )}
-          </div>
-        </div>
-        <div className="absolute bottom-5 lg:bottom-10 right-5 lg:right-10 cursor">
-          {isVideoMuted ? (
-            <button onClick={() => setIsVideoMuted(false)}>
-              <HiVolumeOff className="text-white text-2xl lg:text-4xl" />
-            </button>
-          ) : (
-            <button onClick={() => setIsVideoMuted(true)}>
-              <HiVolumeUp className="text-white text-2xl lg:text-4xl" />
-            </button>
-          )}
-        </div>
-      </div>
+      {post.type === "video" ? (
+        <VideoDetail url={post.video.asset.url} />
+      ) : (
+        <NftDetail url={post.video.asset.url} />
+      )}
 
       <div className="relative w-[1000px] md:w-[900px] lg:w-[700px]">
         <div className="lg:mt-5 mt-2">
